| bug fix for a very slow drop speed (the arduboy.everyXFrames() was taking a char, and the number in parameter was too big, so change it to an int) | bug fix: fix a division by zero when the min drop speed was set to 0 | finally add an EOL after the exported LevelData text

+++ CustomArduboy.cpp
@@ -51,7 +51,7 @@ void CustomArduboy::setFrameRate(uint8_t rate)
   eachFrameMillis = 1000/rate;
 }
 
-bool CustomArduboy::everyXFrames(uint8_t frames)
+bool CustomArduboy::everyXFrames(int frames)
 {
   return frameCount % frames == 0;
 }
